Make `page_script` take effect immediately
Existing issues this aims to resolve:
* Calling `output` multiple times after render will stack up additional
invocations of all `page_*` method calls.
* Calling the canvas `page_*` methods before render does not work when
the canvas is replaced on render start. This happens when the default
paper size and orientation do not match the document.

I think that this behavior is also easier to understand. The `page_*`
methods now need to be called after render or, if used in embedded PHP
code, at the end of the document.